user-location/requirements.txt
flask==3.0.3
gunicorn==22.0.0
redis==5.0.4
pg8000==1.31.2
google-cloud-alloydb-connector[pg8000]==1.1.0

# OpenTelemetry
opentelemetry-sdk==1.24.0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-redis==0.45b0
